Parse daily digest for HR into a JS object.
The daily digest for the House of Representatives was parsed into a JS object. Notable keys included in the object are passedMeasures and failedMeasures. These now have values of an array of JS objects, each of which includes the measure's roll call vote number, its description, and the H-pages the item is located on